% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_rmd.R
\name{run_all_rmd}
\alias{run_all_rmd}
\title{Execution of R code in RMD files}
\usage{
run_all_rmd(filename)
}
\arguments{
\item{filename}{RMD file name}
}
\description{
Execution of R code in RMD files
}
\details{
Convert RMD file in to a R script and execute the R script
The R script created is a temporary file and is not saved
}
